requirements.txt
streamlit>=1.32.0

# YouTube download
yt-dlp>=2024.12.23

# Environment variables
python-dotenv>=1.0.1

# HTTP requests (LLM router, APIs)
requests>=2.31.0

# Google YouTube API (auth, upload, scheduler)
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# LLM providers (as per imports present)
google-generativeai>=0.4.1
openai>=1.12.0
